Image classification + feature extraction/requirements.txt
scikit-learn==1.3.0
Pillow==10.0.0
img2vec-pytorch==1.0.1
torch
torchvision

